fix: restore parseKeyValueLines closing brace + Double/Float literals
The 9f4ef43 MCPAddSheet edit accidentally dropped the closing brace of
parseKeyValueLines, nesting MCPPresetPicker/Row inside it (local
function — 'private' modifier invalid). Restored the brace; Picker and
Row are top-level again. Also: percent=100 → 100.0 (Double field) and
progress bar animation Double → Float.

+++ src/android/app/src/main/java/com/openminis/app/ui/chat/CompactProgressCard.kt
@@ -156,7 +156,7 @@ internal fun CompactProgressCard(
                 }
             } else {
                 val animated by animateFloatAsState(
-                    targetValue = progress.percent / 100.0,
+                    targetValue = (progress.percent / 100.0).toFloat(),
                     animationSpec = tween(220),
                     label = "compactBar",
                 )
